Created LoginStyle in styles package
