+++ STONKS/screens/Login.js
@@ -5,6 +5,7 @@ import {Image, StyleSheet, Text, TextInput, TouchableOpacity, View} from 'react-
 import API_URL from "../url";
 import {UserContext} from "../UserContext";
 import verifyUser from '../request/verifyUser.js';
+import styles from '../styles/LoginStyle';
 
 const Login = (props) => {
 
@@ -99,46 +100,5 @@ const Login = (props) => {
     );
 
 }
-
-/**
- * Styles to make it pretty
- */
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        alignItems: 'center',
-        justifyContent: 'center',
-        backgroundColor: 'white',
-    },
-    button: {
-        alignItems: 'center',
-        backgroundColor: '#e1663b',
-        width: 250,
-        height: 44,
-        padding: 10,
-        borderRadius: 5,
-        marginBottom: 10,
-    },
-    buttonText: {
-        fontSize: 20,
-        alignItems: 'center',
-        justifyContent: 'center',
-        color: 'white',
-        fontWeight: 'bold',
-    },
-    input: {
-        width: 250,
-        fontSize: 18,
-        height: 44,
-        padding: 10,
-        borderWidth: 1,
-        borderColor: 'black',
-        marginVertical: 10,
-    },
-    errorText: {
-        fontSize: 10,
-        color: 'red',
-    },
-});
 
 export default Login;
