+++ STONKS/styles/LoginStyle.js
@@ -1,0 +1,41 @@
+import {StyleSheet} from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+        backgroundColor: 'white',
+    },
+    button: {
+        alignItems: 'center',
+        backgroundColor: '#e1663b',
+        width: 250,
+        height: 44,
+        padding: 10,
+        borderRadius: 5,
+        marginBottom: 10,
+    },
+    buttonText: {
+        fontSize: 20,
+        alignItems: 'center',
+        justifyContent: 'center',
+        color: 'white',
+        fontWeight: 'bold',
+    },
+    input: {
+        width: 250,
+        fontSize: 18,
+        height: 44,
+        padding: 10,
+        borderWidth: 1,
+        borderColor: 'black',
+        marginVertical: 10,
+    },
+    errorText: {
+        fontSize: 10,
+        color: 'red',
+    },
+});
+
+export default styles;
